// src/interfaces/estadisticas.model.ts
export class BusquedaEstadisticas {
    NumPoliza?: string = '';
    NumEndoso?: string = '';
    FecInicio?: string = '';
    FecFin?: string = '';
}

/*export class Estadisticas {
    NumPoliza?: string = '';
    NomProceso?: string = '';
    DescripEstado?: string = '';
    FecRegistro?: string = '';
    FecCierre?: string = '';
    DocIdentidad?: string = '';
    NomEmpresa?: string = '';
    FecEnvioNotif?: string = '';
    FecFirmaDigital?: string = '';
    FecRecepNotif?: string = '';
    FecVisualizaContrato?: string = '';
    NomUsuario?: string = '';
    TipoUsuario?: string = '';
    NumEndoso?: string = '';
}*/